requirements.txt
# requirements
discord.py
